requirements.txt
alembic==1.8.1
anyio==3.6.1
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
distlib==0.3.5
docutils==0.19
FastAPI-SQLAlchemy==0.2.1
filelock==3.7.1
flit==3.7.1
flit-core==3.7.1
greenlet==1.1.2
identify==2.5.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
nodeenv==1.7.0
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette
toml==0.10.2
tomli==2.0.1
tomli-w==1.0.0
typing-extensions==4.3.0
urllib3==1.26.11
virtualenv==20.16.2
Werkzeug==2.2.1
zipp==3.8.1
uvicorn==0.17.6
fastapi
celery
flower
redis
fastapi-mail
twilio
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pyotp
fastapi_versioning
